+++ src/cpx_io/cpx_system/cpx_ap/cpx_ap.py
@@ -220,7 +220,9 @@ class CpxAp(CpxBase):
                 f"the timeout is limited to a minimum of {timeout_ms} ms"
             )
         Logging.logger.info(f"Setting modbus timeout to {timeout_ms} ms")
-        value_to_write = timeout_ms.to_bytes(length=4, byteorder="little")
+        value_to_write = timeout_ms.to_bytes(
+            length=ap_modbus_registers.TIMEOUT.length * 2, byteorder="little"
+        )
         self.write_reg_data(
             value_to_write, ap_modbus_registers.TIMEOUT.register_address
         )